Divided users route and query into two, instructors route and query now has everything that has to do with instructors
